requirements.txt
airtest==1.2.6
altair==4.2.0
apkutils2==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
asttokens==2.0.5
asyncio==3.4.3
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
cigam==0.0.3
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
extra-streamlit-components==0.1.55
facebook-wda==1.4.6
fastjsonschema==2.15.3
fonttools==4.33.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.2
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mss==6.1.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.2
notebook==6.4.12
numpy==1.22.4
opencv-contrib-python==4.5.5.64
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pure-python-adb==0.3.0.dev0
py==1.11.0
pyarrow==8.0.0
pycparser==2.21
pydeck==0.7.1
pydot==1.4.2
pyelftools==0.28
Pygments==2.12.0
PyJWT==2.4.0
Pympler==1.0.1
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-magic==0.4.26
python-xlib==0.31
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywinauto==0.6.3
PyYAML==6.0
pyzmq==23.2.0
requests==2.27.1
retry==0.9.2
rich==12.4.4
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.10.0
streamlit-authenticator==0.2.1
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.2
tqdm==4.64.0
traitlets==5.2.2.post1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
whichcraft==0.6.1
widgetsnbextension==3.6.1
wrapt==1.14.1
xmltodict==0.13.0
zipp==3.8.0
